+++ test/unittest/authVerifierSpec.mjs.js
@@ -335,6 +335,14 @@ describe("AuthVerifier [unittest]", function () {
 				expect(res.dkim[1]?.sdid).to.be.equal("example.org");
 				expect(res.dkim[2]?.result).to.be.equal("SUCCESS");
 				expect(res.dkim[2]?.sdid).to.be.equal("unrelated.org");
+				expect(res.dkim[3]?.result).to.be.equal("PERMFAIL");
+				expect(res.dkim[3]?.result_str).to.be.equal("Invalid (test failure)");
+				expect(res.dkim[4]?.result).to.be.equal("PERMFAIL");
+				expect(res.dkim[4]?.result_str).to.be.equal("Invalid");
+				expect(res.dkim[5]?.result).to.be.equal("PERMFAIL");
+				expect(res.dkim[5]?.result_str).to.be.equal("Invalid");
+				expect(res.dkim[6]?.result).to.be.equal("PERMFAIL");
+				expect(res.dkim[6]?.result_str).to.be.equal("Invalid (test failure signed by unrelated)");
 			});
 			it("With secure signature algorithm", async function () {
 				const message = await createMessageHeader("rfc6376-A.2-arh-valid-a_tag.eml");